Moved config to pkg.
